requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
tqdm>=4.62.0
Pillow>=8.3.0
matplotlib>=3.4.0
pyyaml>=5.4.0
requests>=2.26.0
scikit-learn>=0.24.0
tabulate>=0.8.9
pandas>=1.3.0
